New framegen module
The framegen.py module defines classes and a FrameGenerator utility
function that can generate frames according to some distribution, making
it easier to write the other tools. It's based on the original code
written for the biterrorsim but uses real classes (not only
placeholders), has better names for the classes, and is more modular.